Added error tests
This adds tests to check the errors for the password grant.  At
this point, the password grant implementation is complete and under
full test coverage.  The two errors for password authentication,
dealing with client credentials that are not included and those
which are not valid, have been added and the code has been adjusted
to use them.